+++ crates/sage-cli/src/runner/prefilter.rs
@@ -151,6 +151,8 @@ impl Runner {
             max_fragment_charge: self.parameters.max_fragment_charge,
             wide_window: self.parameters.wide_window,
             min_peaks: self.parameters.min_peaks,
+            min_matched_peaks: db_params.prefilter_min_matched_peaks,
+            max_peaks: db_params.prefilter_max_peaks,
         };
         SpectrumIndexBuilder::new(settings, &db_params.mass_offset_modifications())
     }